+++ tests/test_input_checks.py
@@ -139,50 +139,50 @@ def badInput_rotate5():
 
 
 # BAD BASEGEO .ROTATE_FROM_ANGAX INPUT-----------------------------------------
-def badInput_rotate_from_angax1():
+def badInput_rotate1():
     """ bad ang input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax('123', (1,2,3))
+    bg.rotate('123', (1,2,3))
 
-def badInput_rotate_from_angax2():
+def badInput_rotate2():
     """ bad axis input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, 1)
+    bg.rotate(123, 1)
 
-def badInput_rotate_from_angax3():
+def badInput_rotate3():
     """ bad anchor input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, (1,2,3), anchor=1)
+    bg.rotate(123, (1,2,3), anchor=1)
 
-def badInput_rotate_from_angax4():
+def badInput_rotate4():
     """ bad start input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, (1,2,3), start=1.1)
+    bg.rotate(123, (1,2,3), start=1.1)
 
-def badInput_rotate_from_angax5():
+def badInput_rotate5():
     """ bad increment input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, (1,2,3), increment=None)
+    bg.rotate(123, (1,2,3), increment=None)
 
-def badInput_rotate_from_angax6():
+def badInput_rotate6():
     """ bad degrees input type"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, (1,2,3), degrees=None)
+    bg.rotate(123, (1,2,3), degrees=None)
 
-def badInput_rotate_from_angax7():
+def badInput_rotate7():
     """ bad angle input format"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax([[1,2,3]]*2, (1,2,3))
+    bg.rotate([[1,2,3]]*2, (1,2,3))
 
-def badInput_rotate_from_angax8():
+def badInput_rotate8():
     """ bad axis input format"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax([1,2,3], [1,2,3,4])
+    bg.rotate([1,2,3], [1,2,3,4])
 
-def badInput_rotate_from_angax9():
+def badInput_rotate9():
     """ bad axis input format"""
     bg = BaseGeo(a3, r0)
-    bg.rotate_from_angax(123, [0,0,0])
+    bg.rotate(123, [0,0,0])
 
 
 # EXCITATIONS -----------------------------------------
@@ -317,17 +317,17 @@ class TestExceptions(unittest.TestCase):
         self.assertRaises(MagpylibBadUserInput, badInput_rotate4)
         self.assertRaises(MagpylibBadInputShape, badInput_rotate5)
 
-    def test_rotate_from_angax(self):
-        """ bad .rotate_from_angax inputs"""
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax1)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax2)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax3)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax4)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax5)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax6)
-        self.assertRaises(MagpylibBadInputShape, badInput_rotate_from_angax7)
-        self.assertRaises(MagpylibBadInputShape, badInput_rotate_from_angax8)
-        self.assertRaises(MagpylibBadUserInput, badInput_rotate_from_angax9)
+    def test_rotate(self):
+        """ bad .rotate inputs"""
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate1)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate2)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate3)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate4)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate5)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate6)
+        self.assertRaises(MagpylibBadInputShape, badInput_rotate7)
+        self.assertRaises(MagpylibBadInputShape, badInput_rotate8)
+        self.assertRaises(MagpylibBadUserInput, badInput_rotate9)
 
     def test_magnetization_input(self):
         """ bad magnetization inputs"""