+++ CHANGELOG.md
@@ -84,7 +84,7 @@ All objects (Sensors, Sources, Collections) have additional direct access to
 ### Class method changes
 
 - The class methods `.rotate(angle, axis, anchor)` have been replaced by a new `.rotate(rotation, anchor, increment, start)` method where `rotation` ist a scipy `Rotation` object.
-- The original angle-axis-anchor rotation is now provided by the new method `.rotate_from_angax(angle, axis, anchor, increment, start, degrees)`.
+- The original angle-axis-anchor rotation is now provided by the new method `.rotate(angle, axis, anchor, increment, start, degrees)`.
   - The argument `axis` can now easily be set to the gloabl CS axes with `"x"`, `"y"`, `"z"`.
   - The anchor argument `anchor=0` represents the origin `(0,0,0)`.
   - `angle` argument is in units of deg by default. It can now be set to rad unsing the `degrees` argument.